Review fixes: enforce the schema, bound the offset, pin the meter walk
Three from review on #2092.

**oneOf was advertised and not enforced.** validateValue implements
anyOf and ignores unknown keywords, so transport.seekRelative accepted
both {} and a payload carrying both fields -- the first no-opping as
seekBeats(0) with an ok response, the second silently preferring
deltaBars. Implemented rather than worked around: an exclusivity that
only the schema text knows about is worse than none.

It does not return on success, and that distinction is the whole fix.
anyOf's branches stand for the entire schema, but oneOf sits beside
properties, required and additionalProperties on the same object, so
returning early would have enforced the exclusivity and skipped every
field check next to it. My first attempt did exactly that and the bounds
test caught it. applyIntegerBounds and applyRequestLimits now recurse
through oneOf as well.

**deltaBars overflowed.** It reached barsAndBeats.bars += deltaBars
after unchecked narrowing at three separate edges. seekBars now takes
long long and bounds it to kMaxBarOffset once, so Lua, JSON and OSC all
hand over their own width and none of them narrows; the addition inside
beatsAtBarOffset is done in 64 bits and clamped. The remote schema
carries an explicit range too, so an absurd value is refused at the edge
rather than quietly clamped behind it.

**The meter walk had no test against a real sequence.** The 7/8 case
was exercising the mock's fixed number, so what it pinned was the
facade's delegation. test_transport_seek_juce.cpp builds an Edit with a
time-signature change and seeks across it in both directions.

Writing it corrected me twice. Tracktion counts a 7/8 bar as seven
beats, not 3.5, so my expected values were wrong and the comments and
the manual asserting 3.5 were wrong with them; all of that now says the
bar length is the project's answer rather than naming one.

+++ magda/scripting/MagdaApiLuaBindings.cpp
@@ -893,7 +893,7 @@ int lua_transport_seek_beats(lua_State* L) {
 }
 
 int lua_transport_seek_bars(lua_State* L) {
-    getApi(L)->transport().seekBars(static_cast<int>(luaL_checkinteger(L, 1)));
+    getApi(L)->transport().seekBars(static_cast<long long>(luaL_checkinteger(L, 1)));
     return 0;
 }
